+++ src/Connection.php
@@ -36,7 +36,7 @@ final class Connection extends AbstractPdoConnection
         }
     }
 
-    public function createCommand(string $sql = null, array $params = []): PdoCommandInterface
+    public function createCommand(string $sql = null, array $params = []): Command
     {
         $command = new Command($this);